refactor(metrics-bridge): share bounded policy readers
